[REF] Remove dead/shadowed portal routes; consolidate shared helpers into AccessControlMixin
Dead-route audit cleanup:
- Delete team_staff_portal.view_team (route /my/team was served by
  team_management_portal; this handler was unreachable).
- Move _parse_portal_datetime, _prepare_events_domain, _get_accessible_teams,
  _get_organizations into AccessControlMixin as the single implementation;
  delete the duplicate/shadowed copies from events_portal, timesheets_portal
  and team_staff_portal (resolves a latent import-order override where
  team_staff_portal's staff-only _get_accessible_teams was silently replaced).
- timesheets_portal now inherits AccessControlMixin.
- Fix locale-drift in TestSecurityIntegration.test_01 (assert field name, not
  the English label, since the portal renders fr-CA).

team_staff_portal coverage 72%->97% (dead code removed). Full suite green
(0 failed of 281 tests).
